pyproject.toml
[project]
name = "ml-workflow-object-detection-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fsspec>=2026.1.0",
    "kaggle>=1.8.3",
    "loguru>=0.7.3",
    "onnxruntime>=1.23.2",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.11",
    "typer>=0.21.1",
    "ml_object_detector>=0.1.0",
]

[dependency-groups]
experiments = [
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.8",
    "mlflow>=3.4.0",
    "onnx>=1.17.0",
    "pillow>=12.1.0",
    "polars>=1.37.1",
    "protobuf>=6.33.4",
    "timm>=1.0.24",
    "torch>=2.9.1",
    "transformers>=4.57.5",
    "typer>=0.21.1",
    "ultralytics>=8.4.3",
]

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py311"

[tool.uv.workspace]
members = ["packages/ml-object-detector"]

[tool.uv.sources]
ml-object-detector = { workspace = true }
